+++ AvgScript/src/functions/inlayHint.ts
@@ -3,6 +3,7 @@ import * as vscode from 'vscode';
 import { commandInfoList, inlayHintMap, InlayHintType } from '../lib/dict';
 import { iterateLines } from "../lib/iterateLines";
 import { getAllParams, getCommandType } from '../lib/utilities';
+import { easing_getFuncName, easing_getModeName } from '../lib/easing';
 
 export const inlayHint = vscode.languages.registerInlayHintsProvider('AvgScript', {
     provideInlayHints(document: vscode.TextDocument, range: vscode.Range, token: vscode.CancellationToken) {
@@ -19,7 +20,7 @@ export const inlayHint = vscode.languages.registerInlayHintsProvider('AvgScript'
                     const command = params[0].substring(1);
                     const paramNum = params.length - 1;
                     const paramDefinition = commandInfoList.getValue(command);;
-                    
+
                     let contentStart: number = lineStart + command.length + 1;
 
                     if (paramDefinition === undefined) {
@@ -51,9 +52,15 @@ export const inlayHint = vscode.languages.registerInlayHintsProvider('AvgScript'
 
                         contentStart++;
 
+                        const extraInlayHintInfo = getExtraInlayHintInfo(currentInlayHintType, curParam);
+
                         if (currentInlayHint !== undefined) {
                             let hint = new vscode.InlayHint(new vscode.Position(lineNumber, contentStart)
-                                , currentInlayHint + ":"
+                                , currentInlayHint
+                                + (extraInlayHintInfo === undefined
+                                    ? ''
+                                    : '(' + extraInlayHintInfo + ')')
+                                + ":"
                                 , vscode.InlayHintKind.Parameter);
 
                             hints.push(hint);
@@ -68,3 +75,24 @@ export const inlayHint = vscode.languages.registerInlayHintsProvider('AvgScript'
         return hints;
     }
 });
+
+export const extraInlayHintInfoInvalid = 'Invalid';
+
+export function getExtraInlayHintInfo(type: InlayHintType, param: string) {
+    switch (type) {
+        case InlayHintType.Easing_FuncA:
+        case InlayHintType.Easing_FuncB: {
+            const type = parseInt(param);
+
+            return easing_getFuncName(type);
+        }
+        case InlayHintType.Mode: {
+            const type = parseInt(param);
+
+            return easing_getModeName(type);
+        }
+
+        default:
+            return undefined;
+    }
+}